requirements.txt
typer>=0.12.3
rich>=13.7.1
pydantic>=2.7.3
python-dotenv>=1.0.1
langchain>=0.2.11
langchain-openai>=0.1.17
openai>=1.35.7


